Extend for loops in otemplate by meta variables
In templates it would be useful to compare the current
loop index with a value.

Simply adding a 'loop variable' to the looping dict is problematic
because the sorting order may differ between filling and reading
of the dict.

Problematic example:
 C-Code:
    onion_dict *files = onion_dict_new();
    onion_dict_add(d, "files", files, OD_DICT | OD_FREE_VALUE);
    int loop0=0;
    while ((de = readdir(dir))) {

      onion_dict *file = onion_dict_new();
      onion_dict_add(files, de->d_name, file,
	OD_DUP_KEY | OD_DICT | OD_FREE_VALUE);

      [...]
      char tmp[22];
      onion_dict_add(file, "loop0", tmp, OD_DUP_VALUE);
      loop0 += 1;
    }
    closedir(dir);

 Template:
  {% for file in files %}
    {{ loop0 }}
  {% endfor %}

  Because 'readdir' returns no sorted data, the template loop
  is not 0, 1, … but in random order

This commit adds the meta data variables 'loop0' and 'loop' to for-tags
in templates. They can be used as 'meta.loop0', 'meta.loop' and allows
the detection of the first loop run, ect.
In templates, the variables could only be compared with strings, not numbers.